fix: remove the need to create for of Alpa repo to work with it
this was really stupid idea. Yes the branches will be messy in the
branch Alpa option... but this will cause mess in user space
- in his own fork repo - he have to delete feature branches
or set to delete them automatically which is our job.

So move the mess from user to alpa. Alpa api has to deal with this

But keeping the option for creating packages via fork for
another contributors without commit access - but this is not
currently in the scope of the project goals so it's probably
broken