get rid of gitpython
I thought this would be useful wrapper around git but it
has some bugs (git add) and it often does not have the
features I want to use so I ended up using its direct
command. To end this it just don't run well on small
linux containers like alpine and requires a lot of configuration
to work properly.